# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Haraj Station API
  version: 1.0.0
  description: Generated inventory of route handlers. Request and response schemas are being added per mobile contract.
servers:
  - url: https://harajstation.com
paths:
...
  "/api/identity":
    get:
      summary: "GET /api/identity"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
    post:
      summary: "POST /api/identity"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/listings/{id}":
    patch:
      summary: "PATCH /api/listings/{id}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
  "/api/listings/{id}/comments":
    post:
      summary: "POST /api/listings/{id}/comments"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...